リファクタリング(型名をdefines.hの定義に置き換え)

+++ src/command.h
@@ -1,7 +1,9 @@
 #ifndef _COMMAND_H_INCLUDED_
 #define _COMMAND_H_INCLUDED_
 
-long command_load(void);
-int command_dump(void);
+#include "defines.h"
+
+int32 command_load(void);
+int16 command_dump(void);
 
 #endif
